requirements.txt
asgiref==3.7.2
cloudinary==1.33.0
Django==3.2.20
django-cloudinary-storage==0.3.0
django-filter==23.2
djangorestframework==3.14.0
Pillow==10.0.0
pytz==2023.3
sqlparse==0.4.4
